requirements.txt
beautifulsoup4==4.9.3
Pillow==8.0.1
requests==2.25.0